Dictionary Spacing Bug
Not sure how to solve the problem on the coding end, but being able to copy and paste in a dictionary with hundreds or even thousands of Key names that might be similar seems pretty important. Perhaps putting a check int he Dictionary Editor where if at the very end there is a \n but there is NO space in the textbox, the \n can be deleted from the code?